+++ platform/ios/app/MBXOrnamentsViewController.h
@@ -1,0 +1,5 @@
+#import <UIKit/UIKit.h>
+
+@interface MBXOrnamentsViewController : UIViewController
+
+@end